Rename process_type to calculation_type
