+++ source/databricks/calculation_engine/package/calculation/calculator_args.py
@@ -15,7 +15,7 @@
 from azure.identity import ClientSecretCredential
 from dataclasses import dataclass, field
 from datetime import datetime
-from package.codelists.process_type import CalculationType
+from source.databricks.calculation_engine.package.codelists.calculation_type import CalculationType
 
 
 @dataclass
